requirements.txt
Flask==2.3.2
Flask_HTTPAuth==4.8.0
Flask_JWT_Extended==4.5.2
Flask_Migrate==4.0.4
Flask_RESTful==0.3.10
flask_sqlalchemy==3.0.3
PyJWT==2.7.0
PyJWT==2.7.0
SQLAlchemy==2.0.15
Werkzeug==2.3.4
